Moving MockRegistry out of impl
