+++ codegen/scala-gen/src/test/resources/tests/action-service/with-action-in-name/generated-test-managed/org/example/service/MyServiceActionImplTestKit.scala
@@ -4,9 +4,9 @@ import com.google.protobuf.empty.Empty
 import kalix.scalasdk.Metadata
 import kalix.scalasdk.action.ActionCreationContext
 import kalix.scalasdk.testkit.ActionResult
+import kalix.scalasdk.testkit.TestKitMockRegistry
 import kalix.scalasdk.testkit.impl.ActionResultImpl
 import kalix.scalasdk.testkit.impl.TestKitActionContext
-import kalix.scalasdk.testkit.impl.TestKitMockRegistry
 
 // This code is managed by Kalix tooling.
 // It will be re-generated to reflect any changes to your protobuf definitions.
