+++ sdk/scala-sdk-testkit/src/main/scala/kalix/scalasdk/testkit/TestKitMockRegistry.scala
@@ -14,19 +14,36 @@
  * limitations under the License.
  */
 
-package kalix.scalasdk.testkit.impl
+package kalix.scalasdk.testkit
 
 /**
  * This class is meant to hold mocks used in unit testing cross-component calls
+ *
  * @param mocks
  *   set of mocks or stubs that will be matched by the class upon an external call within a component
  */
-final class TestKitMockRegistry(mocks: Set[Any]) {
-  def get[T](key: Class[T]): Option[T] = mocks.collectFirst {
-    case m if m.getClass.getAnnotatedInterfaces.exists(_.getType == key) => key.cast(m)
-  }
+trait TestKitMockRegistry {
+  def withMock[T](instance: T): TestKitMockRegistry
+
+  private[kalix] def get[T](clazz: Class[T]): Option[T]
 }
 
 object TestKitMockRegistry {
-  val empty = new TestKitMockRegistry(Set.empty)
+  val empty = new TestKitMockRegistryImpl()
+
+  def of(set: Set[Any]): TestKitMockRegistry = new TestKitMockRegistryImpl(set)
+
+  def withMock[T](instance: T): TestKitMockRegistry = new TestKitMockRegistryImpl(Set(instance))
+}
+
+final class TestKitMockRegistryImpl(var mocks: Set[Any] = Set.empty) extends TestKitMockRegistry {
+
+  override def withMock[T](instance: T): TestKitMockRegistry = {
+    mocks = mocks + instance
+    this
+  }
+
+  override def get[T](clazz: Class[T]): Option[T] = mocks.collectFirst {
+    case m if m.getClass.getAnnotatedInterfaces.exists(_.getType == clazz) => clazz.cast(m)
+  }
 }
