+++ sdk/scala-sdk-testkit/src/main/scala/kalix/scalasdk/testkit/impl/TestKitEventSourcedEntityContext.scala
@@ -18,8 +18,7 @@ package kalix.scalasdk.testkit.impl
 
 import akka.stream.Materializer
 import kalix.scalasdk.eventsourcedentity.EventSourcedEntityContext
-import kalix.scalasdk.eventsourcedentity.CommandContext
-import kalix.scalasdk.Metadata
+import kalix.scalasdk.testkit.TestKitMockRegistry
 
 /**
  * INTERNAL API Used by the generated testkit
